+++ functions/app.json
@@ -1,7 +1,7 @@
 {
   "expo": {
-    "name": "huertify",
-    "slug": "huertify",
+    "name": "Huertify.org",
+    "slug": "Huertify.org | agro microescala",
     "version": "1.0.0",
     "orientation": "portrait",
     "icon": "./src/assets/images/Cucumber_leaf.png",
