+++ functions/src/components/EditScreenInfo.tsx
@@ -11,25 +11,18 @@ export default function EditScreenInfo({ texto }: { texto: string }) {
     <View>
       <View style={styles.getStartedContainer}>
         <Text
-          style={styles.getStartedText}
-          lightColor="rgba(0,0,0,0.8)"
-          darkColor="rgba(255,255,255,0.8)">
-          Open up the code for this screen:
+          lightColor={Colors.light.verde}
+          darkColor={Colors.dark.verde}
+          style={styles.getStartedText}>
+            Open up the code for this screen:
         </Text>
 
         <View
-          style={[styles.codeHighlightContainer, styles.homeScreenFilename]}
-          darkColor="rgba(255,255,255,0.05)"
-          lightColor="rgba(0,0,0,0.05)">
-          <MonoText>{texto}</MonoText>
+          style={[styles.codeHighlightContainer, styles.homeScreenFilename]} >
+          <MonoText 
+            lightColor={Colors.light.verde}
+            darkColor={Colors.dark.verde}>{texto}</MonoText>
         </View>
-
-        <Text
-          style={styles.getStartedText}
-          lightColor="rgba(0,0,0,0.8)"
-          darkColor="rgba(255,255,255,0.8)">
-          Change any of the text, save the file, and your app will automatically update.
-        </Text>
       </View>
 
 {/*
